Conflicts:
	src/decaymodes.cc

+++ src/oscaroutput.cc
@@ -206,7 +206,7 @@ void OscarOutput<Format, Contents>::at_interaction(
           file_.get(),
           "# interaction in %zu out %zu rho %12.7f weight %12.7f type %5i \n",
           incoming_particles.size(), outgoing_particles.size(), density,
-          total_cross_section, process_type);
+          total_cross_section, static_cast<int>(process_type));
     } else {
       /* OSCAR line prefix : initial final
        * particle creation: 0 1
@@ -217,7 +217,7 @@ void OscarOutput<Format, Contents>::at_interaction(
        */
       std::fprintf(file_.get(), "%zu %zu %12.7f %12.7f %5i \n",
                    incoming_particles.size(), outgoing_particles.size(),
-                   density, total_cross_section, process_type);
+                   density, total_cross_section, static_cast<int>(process_type));
     }
     for (const auto &p : incoming_particles) {
       write_particledata(p);
@@ -341,6 +341,9 @@ void OscarOutput<Format, Contents>::at_intermediate_time(
    * and at \ref input_general_. **Collision output always gives
    * a list of collisions/decays/box wall crossings plus optionally
    * initial and final configuration.**
+   *
+   * See also \ref collisions_output_in_box_modus_.
+   *
    * Oscar1999
    * ---------
    * Format specifics are the following:\n
